+++ project_config/webpack/loaders/less_loader.js
@@ -2,12 +2,12 @@
  * make less loader
  *
  * @param useCssModules
- * @param cssModulesExclude
+ * @param excludeMatches
  * @param include
  * @param exclude
  * @returns {{test: RegExp, use: *[]}}
  */
-module.exports = (useCssModules = !1, cssModulesExclude = [], include = !1, exclude = !1) => {
+module.exports = (useCssModules = !1, excludeMatches = [], include = !1, exclude = !1) => {
     let loader = {
         test: /\.less$/,
         use: [
@@ -29,8 +29,8 @@ module.exports = (useCssModules = !1, cssModulesExclude = [], include = !1, excl
         ]
     };
 
-    include && (loader.include = cssModulesExclude);
-    exclude && (loader.exclude = cssModulesExclude);
+    include && (loader.include = excludeMatches);
+    exclude && (loader.exclude = excludeMatches);
 
     return loader;
 };