+++ project_config/webpack/util/make_module.js
@@ -20,19 +20,20 @@ module.exports = (config) => {
     var module = {};
     var rules = [babelLoader, htmlLoader];
 
+    let excludeMatches = config.cssModulesExclude;
     if (config.packCssSeparately) {
-        if (config.enableCssModules && config.cssModulesExclude) {
+        if (config.enableCssModules && excludeMatches) {
             rules.push(
-                makeExtractCssLoader(!1, config.cssModulesExclude, !0, !1),
-                makeExtractCssLoader(!0, config.cssModulesExclude, !1, !0),
-                makeExtractLessLoader(!1, config.cssModulesExclude, !0, !1),
-                makeExtractLessLoader(!0, config.cssModulesExclude, !1, !0)
+                makeExtractCssLoader(!1, excludeMatches, !0, !1),
+                makeExtractCssLoader(!0, excludeMatches, !1, !0),
+                makeExtractLessLoader(!1, excludeMatches, !0, !1),
+                makeExtractLessLoader(!0, excludeMatches, !1, !0)
             );
         }
         else if (config.enableCssModules) {
             rules.push(
-                makeExtractCssLoader(!0, config.cssModulesExclude, !1, !1),
-                makeExtractLessLoader(!0, config.cssModulesExclude, !1, !1)
+                makeExtractCssLoader(!0, excludeMatches, !1, !1),
+                makeExtractLessLoader(!0, excludeMatches, !1, !1)
             );
         }
         else {
@@ -43,18 +44,18 @@ module.exports = (config) => {
         }
     }
     else {
-        if (config.enableCssModules && config.cssModulesExclude) {
+        if (config.enableCssModules && excludeMatches) {
             rules.push(
-                makeCssLoader(!1, config.cssModulesExclude, !0, !1),
-                makeCssLoader(!0, config.cssModulesExclude, !1, !0),
-                makeLessLoader(!1, config.cssModulesExclude, !0, !1),
-                makeLessLoader(!0, config.cssModulesExclude, !1, !0)
+                makeCssLoader(!1, excludeMatches, !0, !1),
+                makeCssLoader(!0, excludeMatches, !1, !0),
+                makeLessLoader(!1, excludeMatches, !0, !1),
+                makeLessLoader(!0, excludeMatches, !1, !0)
             );
         }
         else if (config.enableCssModules) {
             rules.push(
-                makeCssLoader(!0, config.cssModulesExclude, !1, !1),
-                makeLessLoader(!0, config.cssModulesExclude, !1, !1)
+                makeCssLoader(!0, excludeMatches, !1, !1),
+                makeLessLoader(!0, excludeMatches, !1, !1)
             );
         }
         else {